requirements.txt
kafka-python
tweepy
praw
torch
transformers
optimum
onnxruntime-gpu
scikit-learn
numpy
nltk
regex
dash
flask
pyspark
matplitlib
plotly
pandas
dash_daq
wordcloud
seaborn
tqdm